Update LCD_Pocket_Pal.ino
Deleted seriously face code, trying to work on "reassured face" frames. Fixed error of making an array of the byte variables, by a stack overflow answer explaining that one should make an array of pointers to those bytes instead. Now the error is that when executing the code, the characters I made aren't printed after the first iteration of the loop. Instead it's all the same unknown character, which could be a result of overlaying the same character in the same slot-however each element for each frame (the eyes, the eyebrows, mouth, etc.) vary from one another in each frame, and the unknown characters being printed are all the same.
This could also have to do with possibly that the lcd object's not clearing the characters I put in previously each time I run "lcd_obj.createChar(0, char_obj);".