Update LCD_Pocket_Pal.ino
Figured out how to assign variables to the items in the 2d array of byte array pointers I made to hold frames. ("reassuring_frames")
Now I have the detail that the LCD is swapping the sides I thought it would write to-I need to edit how I defined the variables and swap the calculations for the sides they're currently being printed to. I also fixed the "raised eyebrow" frames not being mirror images-they weren't the same image, one was a bit higher than the other.